Fix bit shift bugs introduced in r139

+++ trunk/kernel/iscsi.c
@@ -726,7 +726,8 @@ static void set_offset_and_length(struct iet_volume *lu, u8 *cmd, loff_t *off, u
 	case READ_10:
 	case WRITE_10:
 	case WRITE_VERIFY:
-		*off = cmd[2] << 24 | cmd[3] << 16 | cmd[4] << 8 | cmd[5];
+		*off = (u32)cmd[2] << 24 | (u32)cmd[3] << 16 |
+			(u32)cmd[4] << 8 | (u32)cmd[5];
 		*len = (cmd[7] << 8) + cmd[8];
 		break;
 	case READ_16:
@@ -735,7 +736,8 @@ static void set_offset_and_length(struct iet_volume *lu, u8 *cmd, loff_t *off, u
 			(u64)cmd[4] << 40 | (u64)cmd[5] << 32 |
 			(u64)cmd[6] << 24 | (u64)cmd[7] << 16 |
 			(u64)cmd[8] << 8 | (u64)cmd[9];
-		*len = cmd[10] << 24 | cmd[11] << 16 | cmd[12] << 8 | cmd[13];
+		*len = (u32)cmd[10] << 24 | (u32)cmd[11] << 16 |
+			(u32)cmd[12] << 8 | (u32)cmd[13];
 		break;
 	default:
 		BUG();
